revert(utils/moreurls): move parameters type back to moreurls, since it makes no sense to be separate (parameter does make sense though)
