+++ src/search/engines/duckduckgo/search.go
@@ -15,7 +15,6 @@ import (
 	"github.com/hearchco/agent/src/search/scraper/parse"
 	"github.com/hearchco/agent/src/utils/anonymize"
 	"github.com/hearchco/agent/src/utils/moreurls"
-	"github.com/hearchco/agent/src/utils/moreurls/parameters"
 )
 
 func (se Engine) Search(query string, opts options.Options, resChan chan result.ResultScraped) ([]error, bool) {
@@ -95,7 +94,7 @@ func (se Engine) Search(query string, opts options.Options, resChan chan result.
 
 		if pageNum0 == 0 {
 			// Build the parameters.
-			params := parameters.NewParams(
+			params := moreurls.NewParams(
 				paramQueryK, query,
 			)
 
@@ -112,7 +111,7 @@ func (se Engine) Search(query string, opts options.Options, resChan chan result.
 			}
 		} else {
 			// Build the parameters.
-			params := parameters.NewParams(
+			params := moreurls.NewParams(
 				paramQueryK, query,
 				// This value changes depending on how many results were returned on the first page, so it's set to the lowest seen value.
 				paramPageK, strconv.Itoa(pageNum0*20),
