+++ src/search/engines/google/imagesearch.go
@@ -15,7 +15,6 @@ import (
 	"github.com/hearchco/agent/src/search/scraper"
 	"github.com/hearchco/agent/src/utils/anonymize"
 	"github.com/hearchco/agent/src/utils/moreurls"
-	"github.com/hearchco/agent/src/utils/moreurls/parameters"
 )
 
 func (se Engine) ImageSearch(query string, opts options.Options, resChan chan result.ResultScraped) ([]error, bool) {
@@ -106,7 +105,7 @@ func (se Engine) ImageSearch(query string, opts options.Options, resChan chan re
 		ctx.Put("page", strconv.Itoa(i))
 
 		// Build the parameters.
-		params := parameters.NewParams(
+		params := moreurls.NewParams(
 			paramQueryK, query,
 			imgParamTbmK, imgParamTbmV,
 			imgParamAsearchK, imgParamAsearchV,
