+++ src/search/engines/startpage/search.go
@@ -15,7 +15,6 @@ import (
 	"github.com/hearchco/agent/src/search/scraper/parse"
 	"github.com/hearchco/agent/src/utils/anonymize"
 	"github.com/hearchco/agent/src/utils/moreurls"
-	"github.com/hearchco/agent/src/utils/moreurls/parameters"
 )
 
 func (se Engine) Search(query string, opts options.Options, resChan chan result.ResultScraped) ([]error, bool) {
@@ -69,11 +68,11 @@ func (se Engine) Search(query string, opts options.Options, resChan chan result.
 		ctx.Put("page", strconv.Itoa(i))
 
 		// Build the parameters.
-		params := parameters.NewParams(
+		params := moreurls.NewParams(
 			paramQueryK, query,
 		)
 		if pageNum0 > 0 {
-			params = parameters.NewParams(
+			params = moreurls.NewParams(
 				paramQueryK, query,
 				paramPageK, strconv.Itoa(pageNum0+1),
 				paramSafeSearchK, paramSafeSearchV,
